+++ src/components/common/ConnectButton.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { useTranslation } from "next-i18next"
 import React, { useEffect, useState } from "react"
