+++ src/components/ui/UniLink.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import Link from "next/link"
 
 export type UniLinkProps = {
