+++ src/lib/i18n/locales/en/index.json
