+++ src/lib/i18n/locales/ja/common.json
