+++ src/lib/i18n/locales/ja/dashboard.json
