+++ src/lib/i18n/locales/ja/index.json
