+++ src/lib/i18n/locales/zh-TW/dashboard.json
