+++ src/lib/i18n/locales/zh/dashboard.json
