+++ src/lib/i18n/settings.ts
